backend/requirements/requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework==3.15.2
django-filter==24.3
djoser==2.3.1
drf-extra-fields==3.7.0
Pillow==11.0.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-decouple==3.8
python-dotenv==1.0.1
psycopg2==2.9.10